:zap: fix: top 3 en requêtes SQL

+++ src/main/java/com/forkify_backend/api/dto/UserRestaurantStatisticsDto.java
@@ -8,6 +8,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -19,7 +21,7 @@ public class UserRestaurantStatisticsDto {
 
     @NotNull
     @PositiveOrZero
-    private Double totalAmountSpent;
+    private BigDecimal totalAmountSpent;
 
     @NotNull
     @PositiveOrZero
